stop serving students email to client
After students log in their email is served to the client on every
visit. This is completely unnecessary and leaks their email. I've
updated to just return true or false based on login status. The email is
not stored on the client so there should no longer be any security risk
with keeping people logged in. I might even consider ditching email
storage on the backend, and just settle for storing a boolean "was
authenticated" value.