Edit static content for Search Engine Optimization
The title of the site becomes the label on Google, so I've updated it to
be more descriptive. I also added a "description" tag, which Google will
sometimes display on our result. These two changes together should help
identify the purpose of our site when we are found in Google.