feat: hardened the bot to thrown errors
This is one of the steps to get the bot ready for production. Before,
any error would take it offline, but I believe I've managed to wrap most
points of failure in a try/catch that will at least keep the script
running. There's still the possibility that it could be shut down for
good if the state becomes corrupted, but this change should
compartmentalize most errors to only those commands that are effected.