feat: give new users permission to the channel they were invited to
Discord allows you to invite a user directly to a channel, which I've
now coopted to also assign users the proper role to view that channel if
its a class. This is an important feature, as it absolves new users from
having to go through registration.

This is more complicated than it should have been. Discord doesn't
actually tell you which invite a user used to get to the server, but it
does tell you how many times an invite has been used. So I had to save
the use count to the database and do a comparison to figure out which
one was probably used. If this approach finds multiple invites that have
moved by one, it just does nothing, and the user will have to register,
but that will be rare, and isn't any worse than just not having the
feature.