feat: added `/invite` command to create invites
I also consolidated the regex for testing class strings into one at the
top of the file. I needed to extend it to work for upper or lower case,
but I think that makes since, as it makes the code more resiliant. I
also added a trailing "h" or "w" to the regex, for honors or writing
intensive classes respectively.