feat: Add command to send invite links to all class channels
I expect this to be used three times a year to entice new users to
actually invite their classes. The links never expire, and target the
class channel itself, which, given the recent invite detection feature,
should get users directly into the channel itself. The "campaign," as
I'm calling it, happens in two steps. First, the command creates an
ephemeral message to display how the message will look, then a button on
that message triggers the bot to post the messages. This is important,
as the command requires that the admin specifies a template for the
message contents. The template requires `{invite}`, for the invite url,
and allows for `{class}`, for the name and label of the individual
class. Having a confirmation dialogue with an example message will
probably block a few mistakes from going out.