fix: Button emoji's now use discords emoji style, as opposed to systems
I was just inserting the emojis as unicode characters in the text, which
uses whatever browser emoji style the install happened to use. Some
systems have wildly different emoji representations, which was ruining
the point of emojis: to associate buttons with items in the list. Now
that I'm passing the unicode character directly to the emoji option for
buttons, discord seems to be picking it up and styling it as its own.
This also has the added benefit that the emojis are larger and
positioned better.