feat: added command to set role-based or all users visibility of courses
Class Registration will still be created, but it is up to the admins if
they want to display it. This is a stop-gap measure to allow for SR Bot
use in discords with fewer classes, that don't require role-based
visibility. There are only six AST courses the UMN, for instance.

This also fixes a bug where channel permissions were not being updated
in-sync with their categories. That is a fairly major bug that is no
longer an issue.