feat: added `full` option to setup-classes to optionally sort roles
Role sorting is incredibly slow on account of it running no matter if
the order of roles is correct or now. This should allow us to avoid the
slow sorting until it's needed.

Note: It's the discord api calls that are slow; not the in memory
sorting to find the proper order. If I only sorted roles that were out
of order it would be much faster, but I don't have the time at the
moment.