feat: added ability to set up and trigger embed posts from SR Bot
This introduces a bunch of new commands, mostly to manage "files" in the
system, which is how I figured we'd go about managing post contents. The
other option is to use... slash command options, which are terrible for
any actual editing work. I just created a sort of kind-of dsl based on
markdown with headers and splits to create embeds. You can write this
content in a text file, post it, and then "absorb" it with a command.
This will assign it to the database with a name which can then be
referenced later to post the contents as embeds. It's a lot, but it
works. I still want to add the ability to edit prior posts, so that it
doesn't notify people of a new post. I also want to convert @mentions
and #channel-names to their tags: <#1234567890>. I think this will make
things easier to write.