feat: added ability to tag by name either roles or channels in posts
You may now use "{@role}" or "{#channel}" in an absorbed file and SR Bot
will replace it with the proper "<@id>" or "<#id>" tags while posting
the message. This is very useful to avoid having to copy id's all the
time, and allows posts to automatically update if channels or roles are
replaced.