feat: classes may now have a specified category
Discord only allows for 50 categories, which means that if every class
gets its own category, that's a maximum of 50 classes with no other
categories. For some subjects that's just not feasable, so I made it
such that you can specify which category the channel should go into.

I added a new feature to addState. If the state has a default value, it
will automatically become optional for discord.