fix: Allow multiple guilds to add the same class
The classes table was using the class name, math-1234, as the primary
key, but this threw an error if another server tried to add math-1234.
This hasn't come up before as different servers shouldn't be hosting the
same class, but we're trying to split the math server into math classes
and stat classes, and some stat classes broke. The problem is that two
rows with the same primary key can't exist, and the solution, which
should have always been the case, was to make a conjoined primary key
out of the guild and class name. This will require recreating the
classes table (very carefully) in order to make the guild column a
primary key.